Animation: Add skeletal animation data structures and loader integration
Introduce the Libs/Animation module with foundational data types for skeletal
animation (Joint, Skeleton, Skin, AnimationChannel, AnimationClip). Add TRS
conversion utilities (composeTRS, decomposeTRS, toRotationMatrix4, toQuaternion)
with 52 unit tests validating all roundtrip paths between Quaternion, Matrix4,
and CartesianFrame. Update MD5 and GLTF loaders to populate skeleton, skin, and
vertex bone influences/weights into Shape. Data is loaded but not yet consumed
by the runtime (GPU skinning pipeline is the next step).

+++ src/Vulkan/Utility.cpp
@@ -39,7 +39,7 @@ namespace EmEn::Vulkan
 	const char *
 	vkResultToCString (VkResult code) noexcept
 	{
-		switch ( code )
+		switch ( static_cast< int32_t >(code) )
 		{
 			case VK_SUCCESS :
 				return "VK_SUCCESS";
